+++ source/Examples/Resources/ScheduledTask/5-ScheduledTask_CreateScheduledTasksAtLogon_Config.ps1
@@ -34,7 +34,7 @@ Configuration ScheduledTask_CreateScheduledTasksAtLogon_Config
             TaskName           = 'Test task Logon'
             TaskPath           = '\MyTasks'
             ActionExecutable   = 'C:\windows\system32\WindowsPowerShell\v1.0\powershell.exe'
-            ScheduleType       = 'AtLogOn'
+            ScheduleType       = 'AtLogon'
             RepeatInterval     = '00:15:00'
             RepetitionDuration = '08:00:00'
         }